requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-multipart==0.0.20
langchain-groq==0.2.3
langchain==0.3.14
python-dotenv==1.0.1
